% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-files.R
\name{encode_file_presets}
\alias{encode_file_presets}
\title{List or inspect file-selection presets}
\usage{
encode_file_presets(preset = NULL)
}
\arguments{
\item{preset}{Preset name, or \code{NULL} to list all preset names.}
}
\value{
A character vector of preset names or one preset definition.
}
\description{
List or inspect file-selection presets
}
\examples{
encode_file_presets()
encode_file_presets("rnaseq_gene_quant")
}
