% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{encodeUtils-package}
\alias{encodeUtils}
\alias{encodeUtils-package}
\title{encodeUtils: Work with ENCODE metadata and files from R}
\description{
encodeUtils queries ENCODE metadata, lists files attached to experiments,
selects files, previews downloads, downloads chosen files, reads supported
local files, and records citation or manifest metadata.
}
\details{
Start with `encode_search()` for experiments or files. Use `encode_results()`
to extract the displayed table, `encode_list_files()` to list files for
experiments, and `encode_preview_download()` before transferring data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ZohebKhan1/encodeUtils}
  \item \url{https://zohebkhan1.github.io/encodeUtils/}
  \item Report bugs at \url{https://github.com/ZohebKhan1/encodeUtils/issues}
}

}
\author{
\strong{Maintainer}: Zoheb Khan \email{zohebkhan600@gmail.com}

Authors:
\itemize{
  \item Zoheb Khan \email{zohebkhan600@gmail.com}
}

}
