% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{encodeUtils-package}
\alias{encodeUtils}
\alias{encodeUtils-package}
\title{encodeUtils: Work with ENCODE metadata and files from R}
\description{
encodeUtils helps R users find ENCODE datasets and turn ENCODE Portal
metadata into analysis-ready file tables. It queries the ENCODE REST API,
parses JSON returned by experiment and file searches, and standardizes fields
used in scripted analyses, including accessions, assays, biosamples,
organisms, targets, assemblies, output types, file sizes, checksums, and
download URLs.
}
\details{
Use it to search RNA-seq, ChIP-seq, and ATAC-seq experiments, list files from
ENCODE accessions, select common outputs, preview downloads before transfer,
read supported local files into R, and write reproducibility manifests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ZohebKhan1/encodeUtils}
  \item \url{https://zohebkhan1.github.io/encodeUtils/}
  \item Report bugs at \url{https://github.com/ZohebKhan1/encodeUtils/issues}
}

}
\author{
\strong{Maintainer}: Zoheb Khan \email{zohebkhan600@gmail.com} (\href{https://orcid.org/0000-0002-0266-8886}{ORCID})

Authors:
\itemize{
  \item Zoheb Khan \email{zohebkhan600@gmail.com} (\href{https://orcid.org/0000-0002-0266-8886}{ORCID})
}

}
